requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.1
djangorestframework==3.9.0
django-bootstrap-form==3.4
html5lib==1.0.1
Pillow==5.3.0
pycparser==2.19
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2018.5
reportlab==3.5.11
six==1.11.0
tinycss2==0.6.1
WeasyPrint==43
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
